requirements.txt
#dev setup
flake8==6.1.0
black==23.9.0

requests==2.31.0
python-dotenv==1.0.0